requirements.txt
asgiref==3.9.1
Django==5.2.6
django-daisy==1.1.0
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
python-json-logger==3.3.0
sqlparse==0.5.3
django-cors-headers==4.6.0
httpx==0.27.2
mcp==1.1.0
google-generativeai==0.8.3
pgvector==0.4.1
elasticsearch==8.16.0
